Former-commit-id: 32c05338e0530b6ddc98280d46231b025f3c9e86

+++ recovery-nuevoModeloBienesCommon-plugin/src/main/java/es/pfsgroup/plugin/recovery/nuevoModeloBienes/informes/subastabankia/InformeSubastaBean.java
@@ -514,7 +514,7 @@ public class InformeSubastaBean extends InformeSubastaCommon {
 								System.out.println("[INFO] - check lotebien - TipoSubasta70:" + lb.getTipoSubasta70());
 								Float valorTasacion = null;
 								if (!Checks.esNulo(nmbBienLote.getValoraciones()) && nmbBienLote.getValoraciones().size() > 0) {
-									valorTasacion = Checks.esNulo(nmbBienLote.getValoraciones().get(0).getImporteValorTasacion()) ? null : nmbBienLote.getValoraciones().get(0).getImporteValorTasacion();
+									valorTasacion = Checks.esNulo(nmbBienLote.getValoraciones().get(0).getImporteValorTasacion()) ? null : nmbBienLote.getValoraciones().get(0).getImporteValorTasacion().floatValue();
 								}
 								lb.setTasacionActualizada(valorTasacion);
 								System.out.println("[INFO] - check lotebien - Tasacion Actualizada:" + lb.getTasacionActualizada());
